ascii-art-generator/requirements.txt
Pillow>=10.4.0
pyfiglet==1.0.2